webserver/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.31
pika==1.3.2
Flask-SQLAlchemy==3.1.1